backend/requirements.txt
pydantic-ai-slim[openai, graph, logfire]
uvicorn
asyncpg
pydantic-settings
logfire[asyncpg, httpx]
gradio
devtools
rich
python-multipart